add lvgl, fix bugs

+++ sdk/apps/smart_robot/audio/player/tuoyun_flow_player.h
@@ -13,7 +13,7 @@ int tuoyun_audio_write_data(void *buf, int len);
 /* 获取音乐播放器状态  */
 int tuoyun_player_get_status();
 
-void tuoyun_audio_player_set_volume(u8 volume);
+void tuoyun_audio_player_set_volume(s16 volume);
 
 int tuoyun_audio_player_get_volume();
 #endif
